rename to riemann_legacy
