+++ etc/telegraf.conf
@@ -430,8 +430,8 @@
 #   # listen = ":9126"
 
 
-# # Configuration for the Riemann server to send metrics to
-# [[outputs.riemann]]
+# # Configuration for the legacy Riemann plugin
+# [[outputs.riemann_legacy]]
 #   ## URL of server
 #   url = "localhost:5555"
 #   ## transport protocol to use either tcp or udp
